requirements.txt

fastapi
uvicorn
pydantic

transformers
torch
sentencepiece

streamlit
requests

PyPDF2
python-docx